1-249 ETC.tsv (Corrected lots of $'d lines - be more careful)
Went through the first 250 lines, attempting to finish any missing $'d lines, only to find this section was still riddled with errors.  I touched up a handful of stuff while I was at it, but a decent portion of this is error fixes for lines already marked with $'s.  A lot of stuff like this shouldn't pass through review.

------------------------

Line 124: "an herb" is proper for American English.

Line 158: changed "ruins" to "tombs," as it makes more sense in the context of the ambiguity as to when they were placed.
     -Based on the description of the Boowook, it seems like these "defensive creations" were created when Zachariel the Great was still
      around.  That being said, the Royal Mausoleum wouldn't have been ruins at the time of their creation (right?).

Line 162: same as above.

Line 191: I honestly view the plural of Vekarabe to be best kept identical to its singular form.  A knitpicky choice of mine, but correct me if I'm wrong here.

Line 195: Changed "is" to "are," for a similar reason to the above change.  Please correct me if this is not right / you don't agree with this fictional grammar.

Line 200: The description seemed unfinished? I'm not sure, but I went ahead and added a second part the sentence that made it seem way more complete.  Also used an idiom.

Line 216 / 217: Changed "Danderu" into "Dandel" for uniformity.  Please correct me if that was the wrong term to change, but more items and their descriptions are already aligned to the term Dandel.