Synced Monster Names (Need input on a boss name)
Line 2418 / 2419: A pretty funny issue of different names of the same monster.  Sadly I think the item is the one that has to go, as Pawndel is a good name for monster if the other mob is called a Pawnd.  Not sure how line 2417 relates, but I can't even figure out what it drops from (haven't gotten one and can't confirm it drops from a Pawndel, so I'm not touching it).

Line 2429: Changed "Specter Lord" to "Specter Monarch," since that's what the boss is currently titled in other areas of the translation.  I know in the files though, it's referred to as "Specter Master." There are a decent amount of lines referring to this boss as the "Specter Monarch."  I know the wiki calls it "Spector," and it's referenced a couple times as "Spector monarch," but this is honestly a spelling error of specter and is referred to in less than half of the lines, and most of the lines it does appear in already in need of work.  Some input would be nice, and then I could go around changing it, universally.